+++ tests/test_connection.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from paramiko import SFTPError, SSHException
+from paramiko import SFTPError
 from paramiko.ed25519key import Ed25519Key
 
 from common import conn, LOCAL, VFS